Dedup architecture-ratchet SCAN_ROOTS and agent build-script fs helpers
Five test-kernel architecture ratchets hardcoded the same four-root
SCAN_ROOTS array and, in three of them, an identical coverage
sanity-check body; hoist both into repoScan.ts as
ALL_HOST_PRODUCTION_ROOTS and expectRealCoverage().

validate-artifacts.mjs and rewrite-declaration-aliases.mjs each
redefined isFile() and a near-identical recursive directory walker;
extract both into a shared packages/agent/scripts/fsWalk.mjs.

Found and verified by a structural layer-boundary audit; no other
abstraction violations survived scrutiny.
