+++ src/test-kernel/architecture/sessionFactAmbientHelperRetirement.vitest.ts
@@ -6,17 +6,12 @@ import { resolve } from 'node:path';
 import { describe, expect, it } from 'vitest';
 
 import {
+  ALL_HOST_PRODUCTION_ROOTS,
+  expectRealCoverage,
   productionFilesUnder,
   REPO_ROOT,
   stripComments,
 } from '../support/repoScan';
-
-const SCAN_ROOTS = [
-  'packages/cli/src',
-  'packages/desktop/src',
-  'packages/extension/src',
-  'src',
-] as const;
 
 const RETIRED_AMBIENT_HELPER_SYMBOL = /\bemitRuntimeEvent\b/;
 
@@ -27,7 +22,7 @@ function referencesEmitRuntimeEvent(file: string): boolean {
 
 describe('ambient session-fact helper retirement', () => {
   it('keeps the retired ambient session-fact helper out of production code', () => {
-    const references = SCAN_ROOTS.flatMap(productionFilesUnder)
+    const references = ALL_HOST_PRODUCTION_ROOTS.flatMap(productionFilesUnder)
       .filter(referencesEmitRuntimeEvent)
       .toSorted();
 
@@ -35,10 +30,6 @@ describe('ambient session-fact helper retirement', () => {
   });
 
   it('actually scans the production source roots', () => {
-    const scanned = SCAN_ROOTS.reduce(
-      (total, root) => total + productionFilesUnder(root).length,
-      0,
-    );
-    expect(scanned).toBeGreaterThan(100);
+    expectRealCoverage(ALL_HOST_PRODUCTION_ROOTS);
   });
 });
